Functions moved from OSRS_hiscores.py to config package.
